chore: refactor session to mach new project structure.
