+++ pdf_processor_service/main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 from routers import health
 from routers import document
-from routers.sessions import sessions
+from routers import session
 import logging
 
 # Set up logger
@@ -14,4 +14,4 @@ app = FastAPI()
 
 app.include_router(health.router)
 app.include_router(document.router)
-app.include_router(sessions.router)
+app.include_router(session.router)
